refactor: split main file to pages
