+++ .dockerignore
@@ -2,18 +2,24 @@
 .editorconfig
 .env
 .git
+.gitignore
+.github
 .idea
 .vscode
+biome.json
 CHANGELOG.md
+compose.yaml
 coverage*
 data
 docker-compose*
 Dockerfile*
 eslint.config.js
 helm-charts
+images
 LICENSE
 Makefile
 node_modules
 prettier.config.js
 README.md
 renovate.json
+SECURITY.md
