+++ package.json
@@ -32,19 +32,14 @@
     "@tailwindcss/postcss": "^4.1.7",
     "@total-typescript/ts-reset": "^0.6.1",
     "@types/bun": "^1.2.14",
-    "@types/eslint-plugin-tailwindcss": "^3.17.0",
     "@types/node": "^22.15.21",
-    "autoprefixer": "^10.4.21",
-    "cssnano": "^7.0.7",
     "eslint": "^9.27.0",
-    "eslint-plugin-readable-tailwind": "^2.1.2",
+    "eslint-plugin-better-tailwindcss": "^3.0.0",
     "eslint-plugin-simple-import-sort": "^12.1.1",
-    "eslint-plugin-tailwindcss": "4.0.0-alpha.0",
     "globals": "^16.1.0",
     "knip": "^5.57.2",
     "npm-run-all2": "^8.0.3",
     "postcss": "^8.5.3",
-    "postcss-cli": "^11.0.1",
     "prettier": "^3.5.3",
     "tailwind-scrollbar": "^4.0.2",
     "tailwindcss": "^4.1.7",
