+++ src/helpers/env.ts
@@ -1,0 +1,17 @@
+export const ACCOUNT_REGISTRATION =
+  process.env.ACCOUNT_REGISTRATION?.toLowerCase() === "true" || false;
+
+export const HTTP_ALLOWED =
+  process.env.HTTP_ALLOWED?.toLowerCase() === "true" || false;
+
+export const ALLOW_UNAUTHENTICATED =
+  process.env.ALLOW_UNAUTHENTICATED?.toLowerCase() === "true" || false;
+
+export const AUTO_DELETE_EVERY_N_HOURS = process.env.AUTO_DELETE_EVERY_N_HOURS
+  ? Number(process.env.AUTO_DELETE_EVERY_N_HOURS)
+  : 24;
+
+export const HIDE_HISTORY =
+  process.env.HIDE_HISTORY?.toLowerCase() === "true" || false;
+
+export const WEBROOT = process.env.WEBROOT ?? "";
