+++ src/pages/download.tsx
@@ -1,0 +1,66 @@
+
+import { Elysia } from "elysia";
+import { userService } from "./user";
+import { WEBROOT } from "../helpers/env";
+import sanitize from "sanitize-filename";
+import { outputDir } from "..";
+import db from "../db/db";
+
+export const download = new Elysia()
+  .use(userService)
+  .get(
+    "/download/:userId/:jobId/:fileName",
+    async ({ params, jwt, redirect, cookie: { auth } }) => {
+      if (!auth?.value) {
+        return redirect(`${WEBROOT}/login`, 302);
+      }
+
+      const user = await jwt.verify(auth.value);
+      if (!user) {
+        return redirect(`${WEBROOT}/login`, 302);
+      }
+
+      const job = await db
+        .query("SELECT * FROM jobs WHERE user_id = ? AND id = ?")
+        .get(user.id, params.jobId);
+
+      if (!job) {
+        return redirect(`${WEBROOT}/results`, 302);
+      }
+      // parse from url encoded string
+      const userId = decodeURIComponent(params.userId);
+      const jobId = decodeURIComponent(params.jobId);
+      const fileName = sanitize(decodeURIComponent(params.fileName));
+
+      const filePath = `${outputDir}${userId}/${jobId}/${fileName}`;
+      return Bun.file(filePath);
+    },
+  )
+  .get(
+    "/zip/:userId/:jobId",
+    async ({ params, jwt, redirect, cookie: { auth } }) => {
+      // TODO: Implement zip download
+      if (!auth?.value) {
+        return redirect(`${WEBROOT}/login`, 302);
+      }
+
+      const user = await jwt.verify(auth.value);
+      if (!user) {
+        return redirect(`${WEBROOT}/login`, 302);
+      }
+
+      const job = await db
+        .query("SELECT * FROM jobs WHERE user_id = ? AND id = ?")
+        .get(user.id, params.jobId);
+
+      if (!job) {
+        return redirect(`${WEBROOT}/results`, 302);
+      }
+
+      // const userId = decodeURIComponent(params.userId);
+      // const jobId = decodeURIComponent(params.jobId);
+      // const outputPath = `${outputDir}${userId}/`{jobId}/);
+
+      // return Bun.zip(outputPath);
+    },
+  )
